requirements.txt
django==3.0.5
psycopg2==2.8.5
python-dotenv==0.13.0
requests==2.23.0
djangorestframework